+++ IModule.cs
@@ -33,6 +33,6 @@ namespace LibNoise
 
     public interface IModule
     {
-        double GetValue(double x, double y, double z);
+        float GetValue(float x, float y, float z);
     }
 }
